Number counter
Implementation of a number counter in Qt using cmake  to configure libraries to use.

The quit button is a simple push button that uses the plusminus widget as a parent to show on the same window created in main.

The plus minus push buttons and labeled are configured through a grid layout