Added different menu types + skeleton application
Created example classes for various types of menus that can be used for a main window menu. Also created a skeleton main window with a tool bare and menu.

The skeleton has been incorporated into the number counter app where the main window class (skeleton) instantiates a PlusMinus class and makes it a child of the skeleton. While the plus minus class handles the creation of the grid layout the skeleton positions the plusminus widget in the center of the main window